Logo and user added
Made some changes to the landing page to include a logo and a button to
go to the login page.  Also added link to create an account and a user
base class.  Need to correct an error on accountCreation.jsp that should
not display an error on mailing address when the checkbox is checked.
Also need to add servlet to create user.